feat(multi): 多频率域仿真支持与LLM模拟器增强 [core:v1.4.2, tier6:v1.8.5, web:v1.9.5]
【Core 核心仿真引擎】
- 新增多频率域仿真架构：引入 CYCLES_PER_NS、NETWORK_SCALE、IP_SCALE 参数
- 添加 _calculate_time_scale() 自动计算时间缩放因子，支持网络域与IP域独立频率
- 新增 IP_FREQUENCY 配置项，支持 IP 核与网络不同频率运行
- 重构 IP Interface 的 inject_step/eject_step，区分网络域和IP域操作周期
- 所有延迟转换统一使用 CYCLES_PER_NS 替代原 NETWORK_FREQUENCY
- 更新 KCIN/DCIN 配置解析，支持浮点频率与精度校验

【Tier6+ LLM部署分析工具】
- 后端：添加完整配置验证机制（validate_model_config, validate_hardware_config 等）
- 后端：修正 MLA KV Cache 计算公式（基于 DeepSeek-V3 论文 arXiv:2412.19437）
- 后端：KV Cache 维度改为 kv_lora_rank + qk_rope_head_dim（如 512+64=576）
- 后端：优化 MoE 模型参数计算，区分路由专家与共享专家
- 后端：API 增加详细日志记录与异常分类处理
- 前端：App.tsx 大规模重构，引入 WorkbenchContext 状态管理
- 前端：新增 KnowledgeGraph 知识图谱可视化组件
- 前端：侧边栏宽度调整优化（默认520px，最小380px）

【Web 仿真平台】
- 添加完整 KCIN 参数默认值列表（70+ 参数）
- 支持 IP_FREQUENCY 配置与浮点数频率解析
- 优化配置合并逻辑，确保所有参数都有有效默认值

+++ src/analysis/data_collectors.py
@@ -24,7 +24,7 @@ from src.utils.flit import Flit, get_original_source_type, get_original_destinat
 class RequestCollector:
     """请求数据收集器 - 从仿真模型或CSV文件收集请求数据"""
 
-    def __init__(self, network_frequency: float = 2.0):
+    def __init__(self, network_frequency: int = 2):
         """
         初始化请求收集器
