+++ src/analysis/exporters.py
@@ -1647,7 +1647,7 @@ class JSONExporter:
 class ParquetExporter:
     """Parquet格式导出器 - 用于波形数据存储"""
 
-    def __init__(self, network_frequency: float = 2.0):
+    def __init__(self, network_frequency: int = 2):
         """
         初始化Parquet导出器
